create user info modal and add name and email and phone inputs

+++ src/screens/carts/CartScreen.js
@@ -12,13 +12,21 @@ const CartScreen = () => {
     const fetchOrders = async () => {
       try {
         const accessToken = localStorage.getItem('accessToken');
+        if (!accessToken) {
+          setError('No access token found');
+          setLoading(false);
+          return;
+        }
+
         const response = await axios.get(`${process.env.REACT_APP_BACKEND_APP_URL}/api/orders`, {
           headers: {
             Authorization: `Bearer ${accessToken}`,
           },
         });
-        setOrders(response.data);
+
+        setOrders(response.data.data);
       } catch (err) {
+        console.error(err); 
         setError('Failed to fetch orders');
       } finally {
         setLoading(false);
@@ -36,16 +44,21 @@ const CartScreen = () => {
     if (window.confirm("Are you sure you want to delete this order?")) {
       try {
         const accessToken = localStorage.getItem('accessToken');
+        if (!accessToken) {
+          setError('No access token found');
+          return;
+        }
+
         await axios.delete(`${process.env.REACT_APP_BACKEND_APP_URL}/api/orders/${orderId}`, {
           headers: {
             Authorization: `Bearer ${accessToken}`,
           },
         });
 
-        // Update the orders state to remove the deleted order
         setOrders(orders.filter(order => order.id !== orderId));
         alert('Order deleted successfully!');
       } catch (err) {
+        console.error(err); // Log error for debugging
         setError('Failed to delete order');
       }
     }
@@ -62,11 +75,16 @@ const CartScreen = () => {
           orders.map((order) => (
             <div key={order.id} className="bg-white rounded-lg shadow-md overflow-hidden">
               <div className="p-4">
-                <p className="text-gray-600 mb-4">Total Price: ${order.total_price}</p>
+                <p className="text-gray-600 mb-4">Total Price: MAD {order.total_price}</p>
                 <ul className="mb-4">
-                  {order.items.map((item) => (
-                    <li key={item.id} className="text-gray-600">
-                      {item.product.name} (Quantity: {item.quantity}) - ${item.price}
+                  {order.items.map((item, index) => (
+                    <li key={index} className="text-gray-600">
+                      {item.product_name ? (
+                        `${item.product_name}`
+                      ) : (
+                        "Unnamed Product"
+                      )} 
+                      (Quantity: {item.quantity}) - MAD {item.price}
                     </li>
                   ))}
                 </ul>
